+++ test/unit/transaction_builder_test.js
@@ -1,3 +1,4 @@
+import { SorobanDataBuilder } from '../../src/sorobandata_builder.js';
 import { isValidDate } from '../../src/transaction_builder.js';
 const { encodeMuxedAccountToAddress } = StellarBase;
 
@@ -149,6 +150,45 @@ describe('TransactionBuilder', function () {
         .build();
 
       expect(transaction.toEnvelope().v1().tx().ext().switch()).equal(0);
+      done();
+    });
+
+    it("should calculate fee bumps correctly with soroban data", function (done) {
+      sorobanTransactionData = new SorobanDataBuilder().setResourceFee(420).build();
+
+      let transaction = new StellarBase.TransactionBuilder(source, {
+        fee: 620, /* assume BASE_FEE*2 + 420 resource fee */
+        networkPassphrase: StellarBase.Networks.TESTNET,
+      })
+        .addOperation(
+          StellarBase.Operation.invokeHostFunction({
+            func: StellarBase.xdr.HostFunction.hostFunctionTypeInvokeContract(
+              new StellarBase.xdr.InvokeContractArgs({
+                contractAddress: c.address().toScAddress(),
+                functionName: 'test',
+                args: []
+              })
+            ),
+            auth: []
+          })
+        )
+        .setSorobanData(sorobanTransactionData)
+        .setTimeout(StellarBase.TimeoutInfinite)
+        .build();
+
+      expect(
+        transaction.toEnvelope().v1().tx().ext().sorobanData()
+      ).to.deep.equal(sorobanTransactionData);
+
+      const feeBump = StellarBase.TransactionBuilder.buildFeeBumpTransaction(
+        StellarBase.Keypair.random(),
+        "200", // omit resource fee
+        transaction,
+        StellarBase.Networks.TESTNET,
+      );
+
+      expect(feeBump.fee).to.equal("400"); // fee bump is an "op" so double the base
+      sorobanTransactionData = null;
       done();
     });
   });
@@ -363,6 +403,7 @@ describe('TransactionBuilder', function () {
       done();
     });
   });
+
   describe('timebounds', function () {
     it('requires maxTime', function () {
       let source = new StellarBase.Account(
@@ -427,6 +468,7 @@ describe('TransactionBuilder', function () {
       );
     });
   });
+
   describe('setTimeout', function () {
     it('not called', function () {
       let source = new StellarBase.Account(
@@ -569,6 +611,7 @@ describe('TransactionBuilder', function () {
       }).to.not.throw();
     });
   });
+
   describe('.buildFeeBumpTransaction', function () {
     it('builds a fee bump transaction', function (done) {
       const networkPassphrase = 'Standalone Network ; February 2017';